[REF] refactoring pos_order_history_reprint
