+++ pos_orders_history_reprint/models/__init__.py
@@ -1,2 +1,2 @@
 # -*- coding: utf-8 -*-
-from . import models
+from . import model
