+++ pos_orders_history_reprint/static/src/css/main.css
@@ -1,5 +1,3 @@
 .button.reprint {
     left:150px;
 }
-
-
